move sidebar to `/sessions/+layout.svelte`
